Major internal change; switched to Webpack with SASS and Babel
Created package.json to define a bunch of depedencies to be installed with npm.
Removed Bootstrap and Semantic UI CDN's. Bootstrap source code is now installed as a node module. Semantic UI is gone because it can't be use in conjunction with Bootstrap.
All styling will now be written in SASS and all JS will be written in ES6.
Created webpack.config.js to tell webpack how to automate compiling with SASS and Babel.
Moved templates and statics into subdirectories to follow Django standards.